requirements.txt
streamlit>=1.30.0
openai>=1.0.0
groq>=0.4.0
pinecone>=5.0.0
pdfplumber>=0.10.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
